requirements.txt
torch
torchvision
numpy
Pillow
PyYAML>=6.0
